EMP-11102 Failure to serialize response on analytics dispatch no longer retrys already delivered batches.

+++ Exposure/EventSink/Response/AnalyticsConfigResponse.swift
@@ -11,7 +11,7 @@ import Foundation
 /// Configuration data detailing the *Exposure Analytics Environment* for client applications.
 public struct AnalyticsConfigResponse: Decodable {
     /// The requested time untill next contact with the *Analytics Engine*
-    public let secondsUntilNextReport: Int?
+    public let secondsUntilNextReport: Int64?
     
     /// If application metrics should  be included when sending payload
     public let includeApplicationMetrics: Bool?
@@ -26,5 +26,5 @@ public struct AnalyticsConfigResponse: Decodable {
     public let includeDeviceMetrics: Bool?
     
     /// The current timestamp, in unix epoch time.
-    public let timestampNow: Int?
+    public let timestampNow: Int64?
 }
